backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
psycopg[binary]==3.3.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
argon2-cffi==23.1.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
numpy==2.2.2
geopandas==1.0.1
osmnx==2.0.1
